Fixed a crash in MSYS2 when canceling a dialog. Some cleanup as well.
